refactor: create new helper function
- moved the `errorTextGenerator` helper to the `dataErrorMessageGenerator` helper, which we already created
  for error messages, to remove the clutter in the `ImageSelector` component
- added `IMAGE_SELECTOR_ERROR_TYPES` section to `consts.js` to avoid writing keys in switch case with string
- edited the object in the return section of `imagePicker.ts` and added `result.assets[0]` because it contains
  all the information
- changed error messages to be generated inside cases to remove confusion in the `errorTextGenerator` helper
- created a const for the `mimeType` text used in more than one place and added to use this text everywhere
- removed `infoAndErrorText` as it is always an array and extra checking is unnecessary

SVA-1098
