+++ src/components/consul/selectors/ImageSelector.js
@@ -1,4 +1,3 @@
-import * as FileSystem from 'expo-file-system';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
 import { useMutation } from 'react-apollo';
@@ -6,12 +5,7 @@ import { Alert, StyleSheet, TouchableOpacity, View } from 'react-native';
 
 import { colors, consts, Icon, normalize, texts } from '../../../config';
 import { ConsulClient } from '../../../ConsulClient';
-import {
-  deleteArrayItem,
-  formatSize,
-  imageErrorMessageGenerator,
-  jsonParser
-} from '../../../helpers';
+import { deleteArrayItem, errorTextGenerator, jsonParser } from '../../../helpers';
 import { imageHeight, imageWidth } from '../../../helpers/imageHelper';
 import { useSelectImage } from '../../../hooks';
 import { DELETE_IMAGE } from '../../../queries/consul';
@@ -63,44 +57,6 @@ export const ImageSelector = ({ control, errorType, field, imageId, isMultiImage
     onChange(JSON.stringify(imagesAttributes));
   }, [imagesAttributes]);
 
-  const errorTextGenerator = async ({ uri, type }) => {
-    const { size } = await FileSystem.getInfoAsync(uri);
-
-    /* the server does not support files more than 10MB in size. */
-    const volunteerErrorText = size > 10485760 && texts.volunteer.imageGreater10MBError;
-    const consulErrorText = await imageErrorMessageGenerator(uri);
-
-    /* used to specify the mimeType when uploading to the server */
-    const imageType = IMAGE_TYPE_REGEX.exec(uri)[1];
-
-    /* variable to find the name of the image */
-    const uriSplitForImageName = uri.split('/');
-    const imageName = uriSplitForImageName[uriSplitForImageName.length - 1];
-
-    switch (errorType) {
-      case 'consul':
-        setInfoAndErrorText([
-          ...infoAndErrorText,
-          {
-            errorText: texts.consul.startNew[consulErrorText],
-            infoText: `(${type}/${imageType}, ${formatSize(size)})`
-          }
-        ]);
-        break;
-      case 'volunteer':
-        setInfoAndErrorText([
-          ...infoAndErrorText,
-          {
-            errorText: volunteerErrorText,
-            infoText: `${imageName}`
-          }
-        ]);
-        break;
-      default:
-        break;
-    }
-  };
-
   const onDeleteImage = async (index) => {
     if (imageId) {
       try {
@@ -135,10 +91,11 @@ export const ImageSelector = ({ control, errorType, field, imageId, isMultiImage
 
     /* used to specify the mimeType when uploading to the server */
     const imageType = IMAGE_TYPE_REGEX.exec(uri)[1];
+    const mimeType = `${type}/${imageType}`;
 
-    errorTextGenerator({ uri, type });
+    errorTextGenerator({ errorType, infoAndErrorText, mimeType, setInfoAndErrorText, uri });
 
-    setImagesAttributes([...imagesAttributes, { uri, mimeType: `${type}/${imageType}` }]);
+    setImagesAttributes([...imagesAttributes, { uri, mimeType }]);
   };
 
   const values = jsonParser(value);
@@ -202,7 +159,7 @@ export const ImageSelector = ({ control, errorType, field, imageId, isMultiImage
             </TouchableOpacity>
           </WrapperRow>
 
-          {!!infoAndErrorText?.[0]?.infoText && (
+          {!!infoAndErrorText[0]?.infoText && (
             <RegularText smallest>{infoAndErrorText[0].infoText}</RegularText>
           )}
         </>
