+++ src/config/consts.js
@@ -109,6 +109,11 @@ export const consts = {
     3: 3221225472
   },
 
+  IMAGE_SELECTOR_ERROR_TYPES: {
+    CONSUL: 'Consul',
+    VOLUNTEER: 'Volunteer'
+  },
+
   // the image aspect ratio can be overwritten by a global setting `imageAspectRatio`
   // from the server in src/index.js
   IMAGE_ASPECT_RATIO: {
